requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
boto3
openai
Pillow